requirements.txt
# Our requirements
send_mail
mysql.connector